requirements.txt
openai==0.27.7
python-dotenv==1.0.0
tiktoken==0.4.0
numpy==1.24.3
tenacity==8.2.2
pandas==2.0.1